Created new chart filter by type, moved chart to another component
