+++ src/app/admin/dashboard/dashboard.component.html
@@ -33,27 +33,66 @@
       >
         <div class="row">
           <div class="col-xs-12">
-            <img
-              src="/assets/images/chart.png"
-              style="width: 100%">
+            <div
+              class="
+                row
+                align-center
+                justify-center
+                chart-wrapper"
+            >
+              <app-chart
+                class="
+                  chart
+                  row
+                  align-center
+                  justify-center"
+                barWidth="4"
+              ></app-chart>
+            </div>
           </div>
           <div
             class="
               col-xs-12
               col-lg-6"
           >
-            <img
-              src="/assets/images/chart.png"
-              style="width: 100%">
+            <div
+              class="
+                row
+                align-center
+                justify-center
+                chart-wrapper--small"
+            >
+              <app-chart
+                class="
+                  chart
+                  row
+                  align-center
+                  justify-center"
+                barWidth="4"
+              ></app-chart>
+            </div>
           </div>
           <div
             class="
               col-xs-12
               col-lg-6"
           >
-            <img
-              src="/assets/images/chart.png"
-              style="width: 100%">
+            <div
+              class="
+                row
+                align-center
+                justify-center
+                chart-wrapper--small"
+            >
+              <app-chart
+                class="
+                  chart
+                  row
+                  align-center
+                  justify-center"
+                barWidth="4"
+              ></app-chart>
+            </div>
           </div>
         </div>
       </div>
