+++ src/app/admin/dashboard/dashboard.component.scss
@@ -1,0 +1,8 @@
+.chart-wrapper {
+  height: 400px;
+  margin: 40px 0;
+
+  &--small {
+    height: 250px;
+  }
+}
