+++ src/app/advertiser/dashboard/dashboard.component.scss
@@ -5,3 +5,8 @@
     margin: 20px 0;
   }
 }
+
+.chart-wrapper {
+  height: 500px;
+  width: 100%;
+}
