+++ src/app/advertiser/dashboard/dashboard.component.ts
@@ -1,10 +1,9 @@
 import { Component } from '@angular/core';
+import { ChartComponent } from "../../common/components/chart/chart.component";
 
 @Component({
   selector: 'app-dashboard',
   templateUrl: './dashboard.component.html',
   styleUrls: ['./dashboard.component.scss'],
 })
-export class DashboardComponent {
-
-}
+export class DashboardComponent {}
