+++ src/app/common/components/chart-filter-by-type/chart-filter-by-type.component.html
@@ -1,0 +1,66 @@
+<div class="row filters">
+  <div class="col-xs-12 col-sm-6 col-md-3 col-xl-2">
+    <mat-select
+      class="adsh-select"
+      placeholder="All Campaigns"
+      [(value)]="filterData"
+    >
+      <mat-option
+        [value]="'All campaigns'"
+        (click)="filter(value)"
+      >
+        All Campaigns
+      </mat-option>
+      <mat-option
+        [value]="'campaign 1'"
+        (click)="filter(value)"
+      >
+        Campaign 1
+      </mat-option>
+      <mat-option
+        [value]="'campaign 2'"
+        (click)="filter(value)"
+      >
+        Campaign 2
+      </mat-option>
+    </mat-select>
+  </div>
+  <div class="col-xs-12 col-sm-6 col-md-3 col-xl-2">
+    <mat-select
+      class="adsh-select"
+      placeholder="Clicks"
+      [(value)]="filterType"
+    >
+      <mat-option
+        [value]="'clicks'"
+        (click)="filter(value)"
+      >
+        Clicks
+      </mat-option>
+      <mat-option
+        [value]="'impressions'"
+        (click)="filter(value)"
+      >
+        Impressions
+      </mat-option>
+      <mat-option
+        [value]="'ctr'"
+        (click)="filter(value)"
+      >
+        CTR
+      </mat-option>
+      <mat-option
+        [value]="'cpc'"
+        (click)="filter(value)"
+      >
+        Average CPC
+      </mat-option>
+      <mat-option
+        [value]="'cost'"
+        (click)="filter(value)"
+      >
+        Cost
+      </mat-option>
+    </mat-select>
+  </div>
+</div>
