+++ src/app/common/components/chart-filter-by-type/chart-filter-by-type.component.scss
@@ -1,0 +1,7 @@
+.filters {
+  margin: 20px 0;
+
+  > [class*="col-"] {
+    margin: 0 10px;
+  }
+}
