+++ src/app/common/components/chart/chart.component.html
@@ -1,0 +1,15 @@
+<canvas
+  *ngIf="barChartData[0].data.length; else loading"
+  baseChart
+  [datasets]="barChartData"
+  [labels]="barChartLabels"
+  [options]="barChartOptions"
+  [legend]="false"
+  [chartType]="barChartType"
+  [colors]="lineChartColors"
+  (chartHover)="chartHovered($event)"
+  (chartClick)="chartClicked($event)"
+></canvas>
+<ng-template #loading>
+  <mat-spinner [diameter]="30" [strokeWidth]="3" color="primary"></mat-spinner>
+</ng-template>
