+++ src/app/models/initial-state/user.ts
@@ -5,5 +5,5 @@ export const userInitialState: User = {
   email: '',
   isAdvertiser: true,
   isPublisher: true,
-  isAdmin: false,
+  isAdmin: false
 }
