+++ src/app/publisher/dashboard/dashboard.component.html
@@ -29,7 +29,6 @@
       <div class="col-xs-8">
         <app-chart-filter
           *ngIf="!sites"
-          (filter)="getChartData($event)"
         ></app-chart-filter>
       </div>
     </div>
@@ -43,29 +42,120 @@
       >
         <div class="row">
           <div
-            class="col-xs-12"
+            class="col-xs-4"
           >
-            <div>
-              <div style="display: block">
-                <canvas
-                  class="chart"
-                  baseChart
-                  [datasets]="barChartData"
-                  [labels]="barChartLabels"
-                  [options]="barChartOptions"
-                  [legend]="false"
-                  [chartType]="barChartType"
-                  [colors]="lineChartColors"
-                  (chartHover)="chartHovered($event)"
-                  (chartClick)="chartClicked($event)"
-                ></canvas>
-              </div>
+            <div
+              class="
+                row
+                align-center
+                justify-center
+                chart-wrapper"
+            >
+              <app-chart
+                class="
+                  chart
+                  row
+                  align-center
+                  justify-center"
+                barWidth="4"
+              ></app-chart>
             </div>
           </div>
-          <img
-            src="/assets/images/chart.png"
-            style="max-width: 100%;"
+          <div
+            class="col-xs-4"
           >
+            <div
+              class="
+                row
+                align-center
+                justify-center
+                chart-wrapper"
+            >
+              <app-chart
+                class="
+                  chart
+                  row
+                  align-center
+                  justify-center"
+              ></app-chart>
+            </div>
+          </div>
+          <div
+            class="col-xs-4"
+          >
+            <div
+              class="
+                row
+                align-center
+                justify-center
+                chart-wrapper"
+            >
+              <app-chart
+                class="
+                  chart
+                  row
+                  align-center
+                  justify-center"
+              ></app-chart>
+            </div>
+          </div>
+          <div
+            class="col-xs-4"
+          >
+            <div
+              class="
+                row
+                align-center
+                justify-center
+                chart-wrapper"
+            >
+              <app-chart
+                class="
+                  chart
+                  row
+                  align-center
+                  justify-center"
+              ></app-chart>
+            </div>
+          </div>
+          <div
+            class="col-xs-4"
+          >
+            <div
+              class="
+                row
+                align-center
+                justify-center
+                chart-wrapper"
+            >
+              <app-chart
+                class="
+                  chart
+                  row
+                  align-center
+                  justify-center"
+              ></app-chart>
+            </div>
+          </div>
+          <div
+            class="col-xs-4"
+          >
+            <div
+              class="
+                row
+                align-center
+                justify-center
+                chart-wrapper"
+            >
+              <app-chart
+                class="
+                  chart
+                  row
+                  align-center
+                  justify-center"
+              ></app-chart>
+            </div>
+          </div>
         </div>
       </div>
