+++ src/app/publisher/dashboard/dashboard.component.scss
@@ -5,3 +5,9 @@
     margin: 20px 0;
   }
 }
+
+.chart-wrapper {
+  height: 250px;
+  width: 100%;
+}
+
