+++ src/styles/_common.scss
@@ -325,3 +325,8 @@ router-outlet ~ * {
     margin: 0 10px;
   }
 }
+
+.chart {
+  height: 100%;
+  width: 100%;
+}
